refactor(server/studyroom): resolver함수 분리
